Update files after moving to new package.
