+++ src/main/java/com/github/seqware/jobgenerators/PcawgAnnotatorJobGenerator.java
@@ -1,9 +1,11 @@
-package com.github.seqware;
+package com.github.seqware.jobgenerators;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Consumer;
 
+import com.github.seqware.GitUtils;
+import com.github.seqware.OxoGWrapperWorkflow;
 import com.github.seqware.OxoGWrapperWorkflow.Pipeline;
 
 import net.sourceforge.seqware.pipeline.workflowV2.AbstractWorkflowDataModel;
