+++ src/main/java/com/github/seqware/jobgenerators/VariantBamJobGenerator.java
@@ -1,7 +1,10 @@
-package com.github.seqware;
+package com.github.seqware.jobgenerators;
 
 import java.util.Arrays;
 
+import com.github.seqware.GitUtils;
+import com.github.seqware.OxoGWrapperWorkflow;
+import com.github.seqware.TemplateUtils;
 import com.github.seqware.OxoGWrapperWorkflow.BAMType;
 
 import net.sourceforge.seqware.pipeline.workflowV2.AbstractWorkflowDataModel;
